requirements.txt
# Required packages
pygame>=2.5.2
requests>=2.28.0
websocket-client>=1.6.0
